requirements.txt
aniso8601==9.0.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
